+++ battleMenu.cpp
@@ -49,10 +49,10 @@ void BattleMode(StateMachine* sm)
 
 		switch (go)
 		{
-        case List::ADVENTURE: // Search for locations! PROCESS
+        case static_cast<int>(List::ADVENTURE): // Search for locations! PROCESS
             FindLocation(hero, sm);
 			break;
-        case List::TAVERN: // Rest in the tavern! REALIZED
+        case static_cast<int>(List::TAVERN): // Rest in the tavern! REALIZED
             if (hero.getGold() < 150 * hero.getDay())
 				LackOfGoldT(hero);
 			else
@@ -60,16 +60,16 @@ void BattleMode(StateMachine* sm)
             if(hero.getDay() > 30)
                 BadEnd();
 			break;
-        case List::WEAPON: // Weapon shop! REALIZED
+        case static_cast<int>(List::WEAPON): // Weapon shop! REALIZED
 			WeaponShop(hero);
 			break;
-        case List::ITEMS: // Shop with items! REALIZED
+        case static_cast<int>(List::ITEMS): // Shop with items! REALIZED
 			ItemStore(hero);
 			break;
-        case List::INVENTORY: // Player Inventory! REALIZED
+        case static_cast<int>(List::INVENTORY): // Player Inventory! REALIZED
 			Inventory(hero);
 			break;
-        case List::EXIT: // Exit to the main menu! REALIZED
+        case static_cast<int>(List::EXIT): // Exit to the main menu! REALIZED
 			Exit(battle);
 			break;
 		default:
